Refactor rsrch CLI to Commander.js modules
